fix: keep internal consistency with glob and paths
Windows paths threw a spanner in the works. This might not be the
optimal solution for outside modules, in which case I'm open to
changing the behavior to maintain win32 style paths on that OS.
The reasoning here is to keep in line with glob (which recommends
only using forward slashes in patterns) and the URLs which end up
on the Eik server (which, being HTTP URLs, use forward slash).

This should hopefully fix an issue Windows users have with the Eik
CLI not finding the files configured in eik.json